+++ src/app/view-config-upgraders.js
@@ -472,3 +472,15 @@ export function upgradeFrom1_0_13(config) {
     version: '1.0.14',
   };
 }
+
+// Added in version 1.0.15:
+// - Added a fileType-to-dataType mapping
+// so that datasets[].files[].type is no longer required.
+export function upgradeFrom1_0_14(config) {
+  const newConfig = cloneDeep(config);
+
+  return {
+    ...newConfig,
+    version: '1.0.15',
+  };
+}